requirements/tests.txt
factory_boy
pytest
pytest-django>2.7
